+++ lib/rubocop.rb
@@ -156,6 +156,7 @@ require 'rubocop/cop/style/while_until_do'
 require 'rubocop/cop/style/word_array'
 require 'rubocop/cop/style/braces_around_hash_parameters'
 
+require 'rubocop/cop/rails/default_scope'
 require 'rubocop/cop/rails/has_and_belongs_to_many'
 require 'rubocop/cop/rails/output'
 require 'rubocop/cop/rails/read_attribute'